Issue #4437: Modified AvoidStaticImportCheckTest.java and moved its input files to the avoidstaticimport subdirectory
